% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chip_id.R
\name{extract_chip_id}
\alias{extract_chip_id}
\title{Extract numeric identifiers from file names}
\usage{
extract_chip_id(files)
}
\arguments{
\item{files}{Character vector. File names to process.}
}
\value{
A character vector of chip identifiers extracted from the file names.
}
\description{
This function retrieves chip values from file names matching a specific pattern.
}
\examples{
InputPath <- system.file("extdata", package = "ORscrapper")
files <- read_pdf_files(InputPath)

chips <- extract_chip_id(files)

}
