% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORscrapper.R
\name{read_pdf_content}
\alias{read_pdf_content}
\title{Read content from a PDF file}
\usage{
read_pdf_content(file_path)
}
\arguments{
\item{file_path}{Character. The path to the PDF file.}
}
\value{
A character vector, where each element is a line from the PDF content.
}
\description{
This function extracts the text content from a PDF file and splits it into individual lines.
}
