% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdf_files.R
\name{read_pdf_files}
\alias{read_pdf_files}
\title{Read all PDF files in a directory}
\usage{
read_pdf_files(path)
}
\arguments{
\item{path}{Character. Path to the directory to scan for PDF files.}
}
\value{
A character vector with the full paths of the PDF files.
}
\description{
This function scans a specified directory and retrieves all files with a .pdf extension.
}
