% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORscrapper.R
\name{search_ncbi_clinvar}
\alias{search_ncbi_clinvar}
\title{Search for pathogenicity information in NCBI ClinVar}
\usage{
search_ncbi_clinvar(pathogenicity, genes_mutated, total_codifications)
}
\arguments{
\item{pathogenicity}{Ordered list. Existing pathogenicity data.}

\item{genes_mutated}{Ordered list. Existing mutated gene data.}

\item{total_codifications}{Ordered list. Existing mutated gen codification data.}
}
\value{
An updated list of pathogenicity classifications based on NCBI ClinVar search results.
}
\description{
This function queries the NCBI ClinVar database for germline classifications based on gene and codification data.
}
