% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORscrapper.R
\name{extract_values_from_tables}
\alias{extract_values_from_tables}
\title{Extract values from tables within text}
\usage{
extract_values_from_tables(
  lines,
  mutations,
  genes_mutated = list(),
  pathogenicity = list(),
  frequencies = list(),
  codifications = list(),
  changes = list(),
  values = list(),
  start = "Variantes de secuencia de ADN",
  start2 = "   Variaciones del número de copias",
  end = "Genes analizados",
  end2 = "Comentarios adicionales sobre las variantes"
)
}
\arguments{
\item{lines}{Character vector. Lines of text to process.}

\item{mutations}{Character vector. List of known mutation identifiers.}

\item{genes_mutated}{Ordered list to store extracted gene data.}

\item{pathogenicity}{Ordered list to store extracted pathogenicity information.}

\item{frequencies}{Ordered list to store extracted frequency data.}

\item{codifications}{Ordered list to store extracted codification data.}

\item{changes}{Ordered list to store extracted changes data.}

\item{values}{Aggregated list of extracted information.}

\item{start}{Starting marker for the relevant table section.}

\item{start2}{Secondary starting marker for the table section, in case the table is divided in two pages.}

\item{end}{text marker indicating the end of the subset.}

\item{end2}{secondary end marker.}
}
\value{
A list containing extracted data: genes, pathogenicity, frequencies, codifications and changes.
}
\description{
This function analyzes a subset of text lines, extracting information such as mutations, pathogenicity, frequencies, codifications and changes.
}
