% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORscrapper.R
\name{extract_values_start_end}
\alias{extract_values_start_end}
\title{Extract values from start or end patterns}
\usage{
extract_values_start_end(list_input, lines, pattern)
}
\arguments{
\item{list_input}{List. The list to append extracted values to.}

\item{lines}{Character vector. The text lines to search within.}

\item{pattern}{Character. The pattern to search for.}
}
\value{
An updated list with appended values.
}
\description{
This function appends extracted variable values based on start or end markers to a list.
}
