% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pathogenic_only.R
\name{filter_pathogenic_only}
\alias{filter_pathogenic_only}
\title{Filter for pathogenic results only}
\usage{
filter_pathogenic_only(pathogenic_list, related_list)
}
\arguments{
\item{pathogenic_list}{List. A list of pathogenicity classifications.}

\item{related_list}{List. A list of corresponding data to filter alongside pathogenicity.}
}
\value{
A list containing only the elements of the related list corresponding to "Pathogenic" classifications.
}
\description{
This function filters a list of pathogenicity classifications, retaining only those marked as "Pathogenic".
}
